app/requirements.txt
discord.py
pyyaml
sqlalchemy
parse